+++ bika/lims/upgrade/v01_03_006.py
@@ -24,6 +24,7 @@ from bika.lims.catalog.analysis_catalog import CATALOG_ANALYSIS_LISTING
 from bika.lims.config import PROJECTNAME as product
 from bika.lims.upgrade import upgradestep
 from bika.lims.upgrade.utils import UpgradeUtils
+from Products.Archetypes.config import REFERENCE_CATALOG
 
 version = "1.3.6"  # Remember version number in metadata.xml and setup.py
 profile = "profile-{0}:default".format(product)
@@ -55,6 +56,8 @@ def upgrade(tool):
 
     remove_stale_metadata(portal)
 
+    fix_worksheets_analyses(portal)
+
     logger.info("{0} upgraded to version {1}".format(product, version))
     return True
 
@@ -77,3 +80,51 @@ def del_metadata(catalog_id, column):
                     .format(column, catalog_id))
         return
     catalog.delColumn(column)
+
+
+def fix_worksheets_analyses(portal):
+    logger.info("Fix Worksheets Analyses ...")
+    worksheets = portal.worksheets.objectValues()
+    total = len(worksheets)
+    for num, worksheet in enumerate(worksheets):
+        if num and num % 10 == 0:
+            logger.info("Processed worksheets: {}/{}".format(num, total))
+        fix_worksheet_analyses(worksheet)
+    logger.info("Fix Worksheets Analyses [DONE]")
+
+
+def fix_worksheet_analyses(worksheet):
+    """Purge worksheet analyses based on the layout and removes the obsolete
+    relationship from reference_catalog
+    """
+    # Get the referenced analyses via relationship
+    analyses = worksheet.getRefs(relationship="WorksheetAnalysis")
+    analyses_uids = [api.get_uid(an) for an in analyses]
+
+    new_layout = []
+    for slot in worksheet.getLayout():
+        uid = slot.get("analysis_uid")
+        if not api.is_uid(uid):
+            continue
+
+        if uid not in analyses_uids:
+            if is_orphan(uid):
+                continue
+
+        new_layout.append(slot)
+
+    # Re-assign the analyses
+    uids = [slot.get("analysis_uid") for slot in new_layout]
+    worksheet.setAnalyses(uids)
+    worksheet.setLayout(new_layout)
+
+    # Remove all records for this worksheet from reference catalog
+    tool = api.get_tool(REFERENCE_CATALOG)
+    tool.deleteReferences(worksheet, relationship="WorksheetAnalysis")
+
+
+def is_orphan(uid):
+    """Returns whether a counterpart object exists for the given UID
+    """
+    obj = api.get_object_by_uid(uid, None)
+    return obj is None